Former-commit-id: 416d31a643e87bdb3a15abd5ae5cfded5dde2c84

+++ Programs/SceneViewer/Classes/UIControls/TriggerBox.cpp
@@ -27,7 +27,7 @@ bool TriggerBox::AddOption(TriggerBox::OptionID optionId, const DAVA::WideString
         return false;
     }
 
-    int32 textWidth = font->GetStringMetrics(text).width;
+    float32 textWidth = font->GetStringMetrics(text).width;
     float32 buttonWidth = textWidth + 20.f;
     float32 buttonHeight = font->GetSize() * 1.5f;
     Rect buttonRect(nextButtonX, 0.f, buttonWidth, buttonHeight);
